.feature-roomdb created data package and moved model to data package.
