+++ app/src/main/java/com/exercise/matipv2/data/model/MainScreenState.kt
@@ -1,4 +1,4 @@
-package com.exercise.matipv2.model
+package com.exercise.matipv2.data.model
 
 data class MainScreenState(
     var amountInput: String = "",
